requirements.txt
pip
setuptools
wheel
Pillow